modified:   src/librootfs/akaimer39lib/etc/init.d/power_on.sh

+++ src/librootfs/akaimer39lib/etc/init.d/power_on.sh
@@ -1,11 +1,12 @@
 #!/bin/sh
+echo 0 >/sys/class/leds/r_led/brightness
 AP=bus
 Server=10.10.10.2
 wpa_supplicant -B -iwlan0 -Dwext -c /etc/wpa_supplicant.conf
 if iwlist wlan0 scan | grep -i $AP 1>/dev/null ; then
    /etc/init.d/wifi
    Signal=$(iwconfig wlan0 | grep Link | cut -d "=" -f 2 | cut -d "/" -f 1)
-   if [ $Signal -ge 30 ]; then
+   if [ $Signal -ge 30 ] && ssh -i /etc/dropbear/dropbear_rsa_host_key root@$Server 'exit' > /dev/null; then
       if pgrep camera.sh; then killall -9 camera.sh; fi
       dropbear -B
       ntpd -q -p time.windows.com
